Rework page push and pop animations.
Each page can now define its push and pop animation. It might not work in
practice but the beef is there. This also simplifies the PageStack logic a bit
and enables us to push multiple pages without having to check for running animations
but we can still check that via currentPage.popAnimation.running
or currentPage.pushAnimation.running.

opacity behavior is also now enabled. The animation is fast so it does not hurt startup
and it was anyway breaking with the first page.

We also get rid of the legacy unimplemented pageStatus property